requirements.txt
Django==3.0.4
django_heroku==0.3.1
gunicorn==20.0.4
psycopg2==2.7.5
django-webpush==0.3.2
numpy==1.18.4
Pillow==7.1.2